Mattermost publisher becomes the mattermost-live push plugin (step 3)
The router delivers to named push plugins: a subscription's delivery
value resolves in the notice_plugins registry, with at-most-once push
semantics so a push outage never stalls buffered delivery. The
#epicprod-live publication moves into the mattermost-live plugin with
formatting unchanged, selected by the epicprod-live subscription via
new reserved filter keys app_name and live (effective live-policy
state); the importance threshold is the subscription's sublevel list
and the epicprod_live_min_sublevel knob is retired. The
publish_epicprod_live command is now only the tailer loop. Buffered
delivery dedups on (subscriber, dedup_key) at create.
